move example code for spinner
